feat: Implement WandAction.execute and WandConsoleIO
